cores/spi/master: rewrite without fsm
too many cycles were lost in fsm state transitions.

while this was not a concern for low clk frequencies, spi x1 or large
len words, this was becoming an issue for spi x4 and len 8,
with expected spi clk == sys_clk // 2, but actual spi output clk could not reach
more than ~sys_clk // 3 because of the lost cycles.

this is an attempt to fix it, working but not tested widely.