software/usb/device: endpoint should keep reference to the usb device
without reference to the device, and depending on the invocation
the usb device may be collected by the garbage collector while
the usb endpoint is still in use.

this causes the following error which is difficult to trace:
python: core.c:1278: libusb_ref_device: Assertion `refcnt >= 2' failed.

now the endpoint keeps a reference to the device preventing it to be freed